+++ xarray/core/dataarray.py
@@ -5684,7 +5684,7 @@ class DataArray(
         p0: dict[str, Any] | None = None,
         bounds: dict[str, Any] | None = None,
         param_names: Sequence[str] | None = None,
-        kwargs: dict[str, Any] | None = None,
+        **kwargs: Any,
     ) -> Dataset:
         """
         Curve fitting optimization for arbitrary functions.
@@ -5725,7 +5725,7 @@ class DataArray(
             should be manually supplied when fitting a function that takes a variable
             number of parameters.
         **kwargs : optional
-            Additional keyword arguments to passed to scipy curve_fit.
+            Additional keyword arguments passed to scipy curve_fit.
 
         Returns
         -------
@@ -5750,7 +5750,7 @@ class DataArray(
             p0=p0,
             bounds=bounds,
             param_names=param_names,
-            kwargs=kwargs,
+            **kwargs,
         )
 
     def drop_duplicates(